Make `soft` argument of `SoftDeletableModel.delete()` keyword-only
Inserting a new positional argument makes the subclass method
incompatible with the superclass signature. New keyword arguments
do not have this problem.

I chose to require all arguments beyond `using` destined for Django,
such as `keep_parents`, to be passed as keyword arguments as well.
Technically this is incompatible with Django's signature, but the
alternative would be to pass them positionally and risk old code
passing `soft` positionally silently doing the wrong thing:
forwarding the value for `soft` to `keep_parents`.